summary/requirements.txt
fastapi
uvicorn
requests
openpyxl
python-docx
python-pptx
pandas 